Enhance BilgeBlog.Application by adding Slugify.Core package for slug generation, updating CreateCommentCommand and CreatePostCommand to initialize properties, and adding UserId to GetAllPostsQuery and GetPostByIdQuery. Extend PostResult with IsLiked property and update command handlers to handle slug generation and validation. Implement JWT authentication in WebApi with necessary configurations and middleware.

+++ Core/BilgeBlog.Application/Handlers/PostHandlers/Modify/DeletePostCommandHandler.cs
@@ -1,4 +1,5 @@
 using BilgeBlog.Application.DTOs.PostDtos.Commands;
+using BilgeBlog.Application.Exceptions;
 using BilgeBlog.Contract.Abstract;
 using MediatR;
 
@@ -15,6 +16,10 @@ namespace BilgeBlog.Application.Handlers.PostHandlers.Modify
 
         public async Task<bool> Handle(DeletePostCommand request, CancellationToken cancellationToken)
         {
+            var post = await _postRepository.GetByIdAsync(request.Id);
+            if (post == null)
+                throw new NotFoundException("Post", request.Id);
+
             return await _postRepository.RemoveAsync(request.Id);
         }
     }
